refactor: artwork actions render logic (#11159)
* refactor: keep UtilButton as separate component

* refactor: add ArtworkDownloadButton

* refactor: ArtworkActions component

* refactor: ArtworkActions component
